+++ README.md
@@ -1,3 +1,4 @@
 An example project that I learn Golang. Btw, you can use this small Dockerfile repository which using for creating a development for Golang project: https://github.com/duyngha/docker-go
 
-To generate a client run command: `go run cmd/key.go`
+### Commands
+- To generate a client run command: `go run cmd/client/main.go`
